+++ src/main/java/org/dpolivaev/katas/filesystem/domain/internal/memory/SubPage.java
@@ -32,7 +32,7 @@ class SubPage implements Page {
     }
 
     @Override
-    public void write(final long offset, final long length, final byte[] source, final long sourceOffset) {
+    public void write(final long offset, final int length, final byte[] source, final int sourceOffset) {
         this.source.write(start + offset, length, source, sourceOffset);
     }
 
@@ -42,7 +42,7 @@ class SubPage implements Page {
     }
 
     @Override
-    public void read(final long offset, final long length, final byte[] destination, final long destinationOffset) {
+    public void read(final long offset, final int length, final byte[] destination, final int destinationOffset) {
         source.read(offset, length, destination, destinationOffset);
     }
 }
